refactor(gui): split decision API client
